requirements.txt
absl-py==0.10.0
amqp==1.4.9
anyjson==0.3.3
asgiref==3.2.10
astor==0.8.1
backports.entry-points-selectable==1.1.0
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
crispy-forms-gds==0.2.1
cryptography==3.3.1
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.2
Django==3.1.1
django-ajax-datatable==4.3.0
django-angular==2.3
django-bootstrap-datepicker-plus==3.0.5
django-celery==3.3.1
django-cors-headers==3.8.0
django-crispy-forms==1.9.2
django-crontab==0.7.1
django-datatable-view==0.9.0
django-debug-toolbar==3.2.2
django-extensions==3.1.3
django-filter==2.4.0
django-import-export==2.3.0
django-login-required-middleware==0.5.0
django-mptt==0.5.2
django-npm==1.0.0
django-parler==2.2
django-paypal==1.0.0
django-tables2==2.4.0
django-taggit==1.2.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.1
docopt==0.6.2
et-xmlfile==1.0.1
factory-boy==2.12.0
Faker==8.12.1
filelock==3.0.12
gast==0.4.0
grpcio==1.32.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.10
importlib-metadata==4.8.1
isort==5.9.3
itypes==1.2.0
jdcal==1.4.1
Jinja2==3.0.1
joblib==0.17.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kombu==3.0.37
localflavor==1.8
Markdown==3.3.1
MarkupPy==1.14
MarkupSafe==2.0.1
mock==4.0.3
nltk==3.5
npm==0.1.1
odfpy==1.4.1
openpyxl==3.0.5
optional-django==0.1.0
packaging==21.0
pandas==1.1.2
Pillow==8.3.1
pipreqs==0.4.10
platformdirs==2.3.0
pluggy==1.0.0
protobuf==3.13.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
py==1.10.0
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
regex==2020.10.28
requests==2.24.0
rosetta==0.3
scipy==1.5.4
six==1.16.0
sqlparse==0.3.1
tablib==2.0.0
tensorboard==1.12.2
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.2
tox==3.24.3
tqdm==4.51.0
typing-extensions==3.10.0.2
unittest-xml-reporting==3.0.4
uritemplate==3.0.1
urllib3==1.25.10
virtualenv==20.7.2
Werkzeug==1.0.1
whitenoise==5.2.0
xlrd==1.2.0
xlwt==1.3.0
yarg==0.1.9
zipp==3.5.0
